Added in a command to change the calculated origin
Added in the ORG command. ORG is used to calculate the ORiGin offset for a jump, or call immediate instruction. Call register instructions do not get the offset calculation. The advantage with this is that you can write code that can be placed at a random place in memory without writing an on the fly translator or forcing saving to zero or page tables.